+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-1.2.11 (unreleased)
+1.2.11 (2024-05-27)
 -------------------
 
 - WEB-4088 : Add some odwb endpoints (for contacts , for entities)